refactor: S3 패키지 구조 변경
