+++ gusto/src/main/java/com/umc/gusto/global/config/S3Config.java
@@ -1,4 +1,4 @@
-package com.umc.gusto.global.s3;
+package com.umc.gusto.global.config;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
